refactor(git): split the module by subject and colocate tests
